fix(desktop): move isToolResultOnlyMessage to chat-utils to fix Vite HMR Fast Refresh warning
